% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{immoAuth}
\alias{immoAuth}
\title{Reverse Geocoding with the Photon Geocoder for OpenStreetMap, Google Maps, and Bing.}
\source{
https://github.com/mhudecheck/immo/
}
\usage{
immoAuth(key, secret)
}
\arguments{
\item{longitude}{Required. You must enter a valid longitude coordinate;  e.g., -77.0229529}
}
\description{
Enables the use of the Photon geocoder for OpenStreetMap, Google Maps, and Bing to reverse geocode coordinate pairs. Photon allows for unlimited geocode queries, while Google Maps and Bing provide a little more information for 'out of the way' locations. Google Maps and Bing require an API key, and Google Maps limits users to 2,500 free queries a day.
}
\examples{
revgeo(longitude=-77.0229529, latitude=38.89283435)
}
\author{
Michael Hudecheck, \email{michael.hudecheck@gess.ethz.ch}
}
\keyword{api}
\keyword{immo}
\keyword{realestate}
